Split file in two
